Added iteration functionality.
RandomRunner first runs ignoring the iterations entry in the yaml. Then for each spec in the itreactions field, it will run count iterations each time adding the specified offsets (which can be negative) on the corresponding fields of the main configuration. Each path of the yaml that is subject to a iteration spec originates a column in the results csv where that field's value is stored.